Arduino
Project Rotani is an inspection rover project for the Analog Astronaut Training Center Mission Colony 1. This project is based on the 'Cuberover,' a lunar rover designed by Astrobotic. The rover is capable of remote control, rotating the camera 180 degrees left and right, capturing and transmitting photos to a computer, powering on/off, and controlling LEDs for path illumination ahead of the rover. The Rotani rover will be 3D printed and assembled with Arduino components purchased from https://botland.com.pl. The software has been developed using Arduino IDE 2.3.2.
I wholeheartedly recommend https://github.com/Kajetan98 <- working with this guy is such a pleasure and effective experience.